Archeology of the present: Sand Box Stories 2017
For the group final in our Socially Smart Objects and Space class we chose to tell humorous, serious, and empowering stories of the year in 2017. We wanted to show these stories on an interactive platform (Sand Box) that invites an audience to dig and uncover such stories so that they may reflect and take away helpful advise with them into the next year.

Success for us is to see the laughter and enjoyment of our audience play with the sand and reflect on the video projections in a way that leaves them shocked, humored, or pondering the stories presented and the advice exchanged.

The code remarks on the logic of how the resistors were able to read their exposure to light and details the orientation of the sensors in relations to all other sensor values. The projections are visualized in a two by three grid. Projections may need to be flipped depending which side the projector is facing towards the Sand Box.